renamed terraria to tmodloader
