+++ Project-Aurora/Project-Aurora/Profiles/TModLoader/Control_TModLoader.xaml
@@ -1,4 +1,4 @@
-<UserControl x:Class="Aurora.Profiles.Terraria.Control_Terraria"
+<UserControl x:Class="Aurora.Profiles.TModLoader.Control_TModLoader"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -11,12 +11,11 @@
 
     <Grid>
         <TabControl>
-            <TabItem Header="Overview for Terraria">
+            <TabItem Header="Overview for TModLoader">
                 <StackPanel Margin="10">
-                    <CheckBox x:Name="GameEnabled" Content="Enable Aurora to provide lighting effects with Terraria. Ensure you have read the below text." Checked="GameEnabled_Checked" Unchecked="GameEnabled_Checked" />
+                    <CheckBox x:Name="GameEnabled" Content="Enable Aurora to provide lighting effects with TModLoader. Ensure you have read the below text." Checked="GameEnabled_Checked" Unchecked="GameEnabled_Checked" />
                     <TextBlock HorizontalAlignment="Left" Margin="0,10" TextWrapping="Wrap" MaxWidth="824">
                         <Run Text="Support for Terraria requires a mod so that Aurora is able to fetch the variables from the game. This mod requires TModLoader." />
-                        <Run Text="These mods are not authored by the main Aurora developers and any issues with these mods should not be directed at them, but posted on the relevant Github site for the mod." />
                         <Run Text="You can also install the mod through the in-game Mod Browser"/>
                     </TextBlock>
                     <Button Content="TModLoader download" HorizontalAlignment="Left" Margin="0,0,0,0" Padding="10,2" Click="GoToTModLoaderPage_Click" Width="224" />
