+++ Project-Aurora/Project-Aurora/Profiles/TModLoader/GSI/GameState_TModLoader.cs
@@ -1,5 +1,5 @@
 using Aurora.Profiles.Generic.GSI.Nodes;
-using Aurora.Profiles.Terraria.GSI.Nodes;
+using Aurora.Profiles.TModLoader.GSI.Nodes;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
@@ -7,25 +7,25 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Aurora.Profiles.Terraria.GSI {
-    public class GameState_Terraria : GameState<GameState_Terraria> {
+namespace Aurora.Profiles.TModLoader.GSI {
+    public class GameState_TModLoader : GameState<GameState_TModLoader> {
         public ProviderNode Provider => NodeFor<ProviderNode>("provider");
 
         public WorldNode World => NodeFor<WorldNode>("world");
 
         public PlayerNode Player => NodeFor<PlayerNode>("player");
 
-        public GameState_Terraria() : base() { }
+        public GameState_TModLoader() : base() { }
 
         /// <summary>
         /// Creates a GameState_Terraria instance based on the passed JSON data.
         /// </summary>
         /// <param name="JSONstring"></param>
-        public GameState_Terraria(string JSONstring) : base(JSONstring) { }
+        public GameState_TModLoader(string JSONstring) : base(JSONstring) { }
 
         /// <summary>
         /// Creates a GameState_Terraria instance based on the data from the passed GameState instance.
         /// </summary>
-        public GameState_Terraria(IGameState other) : base(other) { }
+        public GameState_TModLoader(IGameState other) : base(other) { }
     }
 }
